A poorly organized but fun data visualization
This was my first attempt at building a GUI for data exploration in MATLAB.  It was built for the Moss Landing Marine Labs Open House 2015 as part of the Physical Oceanography lab demonstrations.  The purpose is to allow users to view CTD data taken from various points around the Monterey Bay.  The GUI consists of two components.  One is a bathymetric contour map that illustrates where each CTD cast was taken.  As this determines both the depth the cast descends to as well as the physical processes one might expect to occur this is highly relevant to interpreting the data.  The second component is the data plot.  Two select inputs allow users to select which cast location and what measurement to plot.  The plots are animated to simulate what oceanographers see when performing a CTD cast and viewing live data.  The color schemes of both windows were chosen to replicate the CTD software used by the scientific crew of the R/V Point Sur, the long-time research vessel